Process git diff content and not working directory
Previously we were finding what files had changed and then reading the
files in the working directory. This was wrong as we want to read what
the file looked like in history and not any working directory
modifications.

Discovered since we modify the gitlab URLs to include credentials when
testing METADATA.

Note: We need to filter with -G which will restricts the diffs to just
show the lines we are interested in (no surrounding lines) and then we
filter again using grep to eliminate diff header information.